# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: gemini
    description: Gemini AI-powered internship application tools
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /gemini/cover-letter:
    post:
      operationId: generateCoverLetter
      tags: [gemini]
      summary: Generate a personalized internship cover letter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CoverLetterInput"
      responses:
        "200":
          description: Generated cover letter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoverLetterOutput"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
        "502":
          description: AI provider request failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CoverLetterInput:
      type: object
      properties:
        candidateName:
          type: string
          minLength: 1
          maxLength: 100
        skills:
          type: array
          items:
            type: string
            maxLength: 60
          minItems: 1
          maxItems: 12
        experience:
          type: string
          maxLength: 1200
        role:
          type: string
          minLength: 1
          maxLength: 160
        company:
          type: string
          minLength: 1
          maxLength: 160
        location:
          type: string
          maxLength: 120
        duration:
          type: string
          maxLength: 80
        description:
          type: string
          maxLength: 1200
      required:
        - candidateName
        - skills
        - role
        - company
    CoverLetterOutput:
      type: object
      properties:
        coverLetter:
          type: string
      required:
        - coverLetter
    GeminiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error