# lib/api-zod/src/generated/types/coverLetterInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CoverLetterInput {
  /**
     * @minLength 1
     * @maxLength 100
     */
  candidateName: string;
  /**
     * @minItems 1
     * @maxItems 12
     * @items.maxLength 60
     */
  skills: string[];
  /** @maxLength 1200 */
  experience?: string;
  /**
     * @minLength 1
     * @maxLength 160
     */
  role: string;
  /**
     * @minLength 1
     * @maxLength 160
     */
  company: string;
  /** @maxLength 120 */
  location?: string;
  /** @maxLength 80 */
  duration?: string;
  /** @maxLength 1200 */
  description?: string;
}